refactor validation
